# colexec: benchmark and optimize window functions