… pkg/sql/colexec/colexecwindow/lag.eg.go
… pkg/sql/colexec/colexecwindow/lead.eg.go