**colexec: add benchmarks for window functions**

This patch adds a benchmark which runs each window function with each
combination of partitioning vs not partitioning, and ordering vs not
ordering. For window functions that take arguments, the argument
columns are simple `int64` columns.

Release note: None

**colexec: optimize window functions for frame with many partitions**

This patch optimizes `SpillingBuffer` and the `lag` and `lead` window
functions for the case when the input is split into many partitions.
This is primarily accomplished in two ways:
1. By removing the need to take a window into input batches for
   `SpillingBuffer.AppendTuples` in the in-memory case.
2. By holding on to memory after calls to `SpillingBuffer.Reset`.
   `lag` and `lead` reset the `SpillingBuffer` at the start of
   each partition, so deeply reseting the `AppendOnlyBufferedBatch`
   is expensive when there are many partitions.

Release note: None