@DrewKimball 2021-07-01
Benchmark results for lead (lag will be similar, and other window functions will not see any change):

**No Optimizations**
```
I210630 23:59:50.437820 1 (gostd) rand.go:92  [-] 1  random seed: -5522927749634229162
goos: darwin
goarch: amd64
pkg: github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecwindow
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkWindowFunctions/LEAD/rows=4096/partition=true/order=false-12         	    1000	    786241 ns/op
BenchmarkWindowFunctions/LEAD/rows=4096/partition=false/order=false-12        	    1000	    142404 ns/op
BenchmarkWindowFunctions/LEAD/rows=32768/partition=true/order=false-12        	    1000	   6162796 ns/op
BenchmarkWindowFunctions/LEAD/rows=32768/partition=false/order=false-12       	    1000	   1167114 ns/op
PASS
ok  	github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecwindow	11.465s
```
**With Optimizations**
```
I210630 23:58:04.594915 1 (gostd) rand.go:92  [-] 1  random seed: -1155557353335678116
goos: darwin
goarch: amd64
pkg: github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecwindow
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkWindowFunctions/LEAD/rows=4096/partition=true/order=false-12         	    1000	    392238 ns/op
BenchmarkWindowFunctions/LEAD/rows=4096/partition=false/order=false-12        	    1000	    129586 ns/op
BenchmarkWindowFunctions/LEAD/rows=32768/partition=true/order=false-12        	    1000	   3063392 ns/op
BenchmarkWindowFunctions/LEAD/rows=32768/partition=false/order=false-12       	    1000	   1094757 ns/op
PASS
ok  	github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecwindow	6.871s
```