@mgartner 2021-07-07
@DrewKimball you might find the [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) tool helpful for comparing benchmarks.

You can install it with `go get golang.org/x/perf/cmd/benchstat`. You'll have to run each benchmark five times with `GOFLAGS='-count=5'` and save the output to a file. Here's an example of benchstat's output:

```
name                                                            old time/op  new time/op  delta
WindowFunctions/LEAD/rows=4096/partition=true/order=false-12     786µs ± 0%   392µs ± 0%  -50.11%  (p=0.008 n=5+5)
WindowFunctions/LEAD/rows=4096/partition=false/order=false-12    142µs ± 0%   130µs ± 0%   -9.00%  (p=0.008 n=5+5)
WindowFunctions/LEAD/rows=32768/partition=true/order=false-12   6.16ms ± 0%  3.06ms ± 0%  -50.29%  (p=0.008 n=5+5)
WindowFunctions/LEAD/rows=32768/partition=false/order=false-12  1.17ms ± 0%  1.09ms ± 0%   -6.20%  (p=0.008 n=5+5)
```