@DrewKimball 2021-07-14
Definitely saving that script, too.

> I'm curious about the choice of `1` for `matchLen` parameter: the input rows will be ordered on `arg1` column which is also constantly increasing. I feel like we would want `partitionCol` to be incremented too, no?

`partitionCol` is just a boolean column to indicate where the partition boundaries are - there aren't actually any concrete values.